Added: Float Number

+++ src/Generics/Generics.java
@@ -2,8 +2,8 @@ package Generics;
 
 public class Generics {
 
-    public int maximum(Integer x, Integer y, Integer z) {
-        Integer max = x;
+    public Float maximum(Float x, Float y, Float z) {
+        Float max = x;
         if (y.compareTo(max) > 0 && y.compareTo(z) > 0)
             max = y;
         else if (z.compareTo(max) > 0)
